refactor(mixtral): remove `tiny` name_prefix
