+++ example/mixtral/checkpoint_loader.h
@@ -9,7 +9,7 @@ namespace infini_train::nn {
 class TransformerModel;
 } // namespace infini_train::nn
 
-namespace tiny_mixtral {
+namespace mixtral {
 
 infini_train::nn::TransformerConfig ConfigFromLLMC(const std::string &filepath);
 
@@ -18,4 +18,4 @@ void CheckLLMCConfig(const std::string &filepath, const infini_train::nn::Transf
 std::shared_ptr<infini_train::nn::TransformerModel>
 LoadFromLLMC(const std::string &filepath, const infini_train::nn::TransformerConfig &expected_config);
 
-} // namespace tiny_mixtral
+} // namespace mixtral
